+++ tournament/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
-from .models import Tournament, Team
+from .models import Tournament, Team, TeamMember
 
 
 admin.site.register(Tournament)
 admin.site.register(Team)
+admin.site.register(TeamMember)